backend/chatbot/pyproject.toml
[tool.poetry]
name = "health-rag"
version = "0.1.0"
description = ""
authors = ["ingeniom22 <jamesmichael0444@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
langserve = "^0.2.0"
langchain = "^0.2.0"
langchain-community = "^0.2.0"
openai = "^1.30.1"
sse-starlette = "^2.1.0"
python-dotenv = "^1.0.1"
qdrant-client = "^1.9.1"
beautifulsoup4 = "^4.12.3"
tiktoken = "^0.7.0"
langchainhub = "^0.1.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
